--- http4k-testing/webdriver/src/test/kotlin/org/http4k/webdriver/assert.kt
@@ -1,6 +1,0 @@
-package org.http4k.webdriver
-
-import com.natpryce.hamkrest.assertion.assertThat
-import com.natpryce.hamkrest.throws
-
-fun isNotImplemented(fn: () -> Unit) = assertThat({ fn() }, throws<FeatureNotImplementedYet>())
